+++ src/main/java/edu/sdsu/its/API/Followup.java
@@ -1,7 +1,8 @@
-package edu.sdsu.its.fit_welcome;
+package edu.sdsu.its.API;
 
 import com.google.gson.Gson;
-import edu.sdsu.its.fit_welcome.Models.User;
+import edu.sdsu.its.Welcome.DB;
+import edu.sdsu.its.API.Models.User;
 import org.apache.log4j.Logger;
 
 import javax.ws.rs.*;
