+++ src/main/java/edu/sdsu/its/API/Models/Event.java
@@ -1,6 +1,6 @@
-package edu.sdsu.its.fit_welcome.Models;
+package edu.sdsu.its.API.Models;
 
-import edu.sdsu.its.fit_welcome.DB;
+import edu.sdsu.its.Welcome.DB;
 import org.joda.time.DateTime;
 
 import java.sql.Timestamp;
