+++ src/main/java/edu/sdsu/its/API/Models/TimeEntry.java
@@ -1,4 +1,4 @@
-package edu.sdsu.its.fit_welcome.Models;
+package edu.sdsu.its.API.Models;
 
 import org.apache.log4j.Logger;
