+++ src/main/java/edu/sdsu/its/API/Models/User.java
@@ -1,6 +1,6 @@
-package edu.sdsu.its.fit_welcome.Models;
+package edu.sdsu.its.API.Models;
 
-import edu.sdsu.its.fit_welcome.DB;
+import edu.sdsu.its.Welcome.DB;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
