+++ src/main/java/edu/sdsu/its/Jobs/SyncUserDB.java
@@ -2,7 +2,7 @@ package edu.sdsu.its.Jobs;
 
 import edu.sdsu.its.Blackboard.Models.User;
 import edu.sdsu.its.Blackboard.Users;
-import edu.sdsu.its.fit_welcome.DB;
+import edu.sdsu.its.Welcome.DB;
 import org.apache.log4j.Logger;
 import org.quartz.*;
