+++ src/main/java/edu/sdsu/its/Welcome/Timesheet.java
@@ -1,6 +1,7 @@
-package edu.sdsu.its.fit_welcome;
+package edu.sdsu.its.Welcome;
 
 import com.opencsv.CSVWriter;
+import edu.sdsu.its.API.Clock;
 import org.apache.log4j.Logger;
 import org.joda.time.format.DateTimeFormat;
